+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 2.29.0
 ### Feature
 * add animation to `exec` and `execAnimation` option [#679](https://github.com/jcubic/jquery.terminal/issues/679)
+* add warning when calling invoke_key on disabled terminal [#678](https://github.com/jcubic/jquery.terminal/issues/678)
 ### Bugfix
 * tweak types for cmd function prompt